FIXED THE BUG!
Ok so basically when you play too fast, the timeouts don't expire and will
fire in the correct phase (but in a later round), causing the timer to jump.
The correct way to handle this is to actually clearTimeout instead of relying
on inexhaustive checks.

Also cleaned up a few things in the client that were logging errors.